src/requirements.txt
# Flask
fastapi
uvicorn[standard]
# asgiref
gunicorn
sqlalchemy[asyncio]
asyncpg
# asyncio
alembic

# PyMySQL
coloredlogs
python-dotenv
